terrain_generation/wave_function_collapse/pyproject.toml
[project]
name = "wave-function-collapse"
version = "0.1.0"
description = "Python implementation of the Wave function collapse (WFC) algorithm."
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "openpyxl>=3.1.5",
    "pillow>=12.2.0",
    "pydantic>=2.13.2",
    "pygame>=2.6.1",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "ruff>=0.15.11",
    "ty>=0.0.31",
]